refactor(agent): generalize harness integrations crate
