+++ packages/browseros-agent/crates/harness-integrations/src/catalog.rs
@@ -1,0 +1,680 @@
+use std::path::PathBuf;
+
+use crate::{
+    AgentId, AgentInfo, AgentScope, Error, McpTransport,
+    paths::{any_exists, pick_config_path, selected_os_paths},
+};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct PerOsPaths {
+    pub darwin: &'static [&'static str],
+    pub linux: &'static [&'static str],
+    pub windows: &'static [&'static str],
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ConfigFormat {
+    Json,
+    Jsonc,
+    Toml,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum KeyTransform {
+    SimpleName,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum InjectValue {
+    String(&'static str),
+    Bool(bool),
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct StdioShape {
+    pub top_level_key: &'static str,
+    pub command_field: Option<&'static str>,
+    pub args_field: Option<&'static str>,
+    pub env_field: Option<&'static str>,
+    pub command_as_array: bool,
+    pub tag_key: Option<&'static str>,
+    pub tag_value: Option<&'static str>,
+    pub injects: &'static [(&'static str, InjectValue)],
+    pub key_transform: Option<KeyTransform>,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct HttpShape {
+    pub url_field: Option<&'static str>,
+    pub header_field: Option<&'static str>,
+    pub tag_key: Option<&'static str>,
+    pub tag_value: Option<&'static str>,
+    pub sse_tag_value: Option<&'static str>,
+    pub injects: &'static [(&'static str, InjectValue)],
+    pub supports_oauth: bool,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ProjectSurface {
+    pub stdio: StdioShape,
+    pub http: Option<HttpShape>,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct McpSurfaceSources {
+    pub first_party: &'static str,
+    pub smithery: Option<&'static str>,
+    pub notes: Option<&'static str>,
+    pub verified: &'static str,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct HarnessDefinition {
+    pub id: AgentId,
+    pub display_name: &'static str,
+    pub install_check_paths: PerOsPaths,
+    pub mcp: McpSurface,
+    pub skill: Option<SkillSurface>,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct McpSurface {
+    pub system_paths: PerOsPaths,
+    pub project_file: Option<&'static str>,
+    pub format: ConfigFormat,
+    pub system_transports: &'static [McpTransport],
+    pub project_transports: Option<&'static [McpTransport]>,
+    pub stdio: StdioShape,
+    pub http: Option<HttpShape>,
+    pub project: Option<ProjectSurface>,
+    pub sources: McpSurfaceSources,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct SkillSurface {
+    pub global_roots: PerOsPaths,
+    pub first_party: &'static str,
+    pub verified: &'static str,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct AgentSurface {
+    pub harness: &'static HarnessDefinition,
+    pub mcp: &'static McpSurface,
+    pub supported_transports: &'static [McpTransport],
+    pub stdio: StdioShape,
+    pub http: Option<HttpShape>,
+}
+
+const SMITHERY_URL: &str = "https://github.com/smithery-ai/cli/blob/main/src/config/clients.ts";
+const VERIFIED: &str = "2026-07-06";
+const NONE: &[(&str, InjectValue)] = &[];
+const ZED_INJECTS: &[(&str, InjectValue)] = &[
+    ("source", InjectValue::String("custom")),
+    ("enabled", InjectValue::Bool(true)),
+];
+const OPENCODE_STDIO_INJECTS: &[(&str, InjectValue)] = &[
+    ("type", InjectValue::String("local")),
+    ("enabled", InjectValue::Bool(true)),
+];
+const OPENCODE_HTTP_INJECTS: &[(&str, InjectValue)] = &[("enabled", InjectValue::Bool(true))];
+
+const STDIO_SSE_HTTP: &[McpTransport] =
+    &[McpTransport::Stdio, McpTransport::Sse, McpTransport::Http];
+const STDIO_HTTP: &[McpTransport] = &[McpTransport::Stdio, McpTransport::Http];
+const STDIO_ONLY: &[McpTransport] = &[McpTransport::Stdio];
+
+const fn paths(
+    darwin: &'static [&'static str],
+    linux: &'static [&'static str],
+    windows: &'static [&'static str],
+) -> PerOsPaths {
+    PerOsPaths {
+        darwin,
+        linux,
+        windows,
+    }
+}
+
+const fn stdio(top_level_key: &'static str) -> StdioShape {
+    StdioShape {
+        top_level_key,
+        command_field: None,
+        args_field: None,
+        env_field: None,
+        command_as_array: false,
+        tag_key: None,
+        tag_value: None,
+        injects: NONE,
+        key_transform: None,
+    }
+}
+
+const fn skill(
+    global_roots: PerOsPaths,
+    first_party: &'static str,
+    verified: &'static str,
+) -> Option<SkillSurface> {
+    Some(SkillSurface {
+        global_roots,
+        first_party,
+        verified,
+    })
+}
+
+const CLAUDE_CODE: HarnessDefinition = HarnessDefinition {
+    id: AgentId::ClaudeCode,
+    display_name: "Claude Code",
+    install_check_paths: paths(
+        &["$HOME/.claude"],
+        &["$HOME/.claude"],
+        &["$USERPROFILE\\.claude"],
+    ),
+    mcp: McpSurface {
+        system_paths: paths(
+            &["$CLAUDE_CONFIG_DIR/.claude.json", "$HOME/.claude.json"],
+            &["$CLAUDE_CONFIG_DIR/.claude.json", "$HOME/.claude.json"],
+            &[
+                "$CLAUDE_CONFIG_DIR\\.claude.json",
+                "$USERPROFILE\\.claude.json",
+            ],
+        ),
+        project_file: Some(".mcp.json"),
+        format: ConfigFormat::Json,
+        system_transports: STDIO_SSE_HTTP,
+        project_transports: Some(STDIO_ONLY),
+        stdio: stdio("mcpServers"),
+        http: Some(HttpShape {
+            url_field: None,
+            header_field: None,
+            tag_key: Some("type"),
+            tag_value: Some("http"),
+            sse_tag_value: Some("sse"),
+            injects: NONE,
+            supports_oauth: true,
+        }),
+        project: Some(ProjectSurface {
+            stdio: StdioShape {
+                tag_key: Some("type"),
+                tag_value: Some("stdio"),
+                ..stdio("mcpServers")
+            },
+            http: None,
+        }),
+        sources: McpSurfaceSources {
+            first_party: "https://docs.claude.com/en/docs/claude-code/mcp",
+            smithery: Some(SMITHERY_URL),
+            notes: Some(
+                "HTTP and SSE entries in ~/.claude.json require an explicit `type` field (\"http\", \"sse\", or \"ws\") or Claude Code emits a \"url but no type\" parse warning and skips the entry on launch. Stdio entries are accepted with or without a type tag. Project scope (.mcp.json) writes an explicit type: stdio tag per Claude Code project-scope docs.",
+            ),
+            verified: "2026-07-10",
+        },
+    },
+    skill: skill(
+        paths(
+            &["$CLAUDE_CONFIG_DIR/skills", "$HOME/.claude/skills"],
+            &["$CLAUDE_CONFIG_DIR/skills", "$HOME/.claude/skills"],
+            &[
+                "$CLAUDE_CONFIG_DIR\\skills",
+                "$USERPROFILE\\.claude\\skills",
+            ],
+        ),
+        "https://code.claude.com/docs/en/skills",
+        "2026-07-22",
+    ),
+};
+
+const CURSOR: HarnessDefinition = HarnessDefinition {
+    id: AgentId::Cursor,
+    display_name: "Cursor",
+    install_check_paths: paths(
+        &["/Applications/Cursor.app"],
+        &["$HOME/.config/Cursor"],
+        &["$APPDATA\\Cursor"],
+    ),
+    mcp: McpSurface {
+        system_paths: paths(
+            &["$HOME/.cursor/mcp.json"],
+            &["$HOME/.cursor/mcp.json"],
+            &["$USERPROFILE\\.cursor\\mcp.json"],
+        ),
+        project_file: Some(".cursor/mcp.json"),
+        format: ConfigFormat::Json,
+        system_transports: STDIO_SSE_HTTP,
+        project_transports: Some(STDIO_SSE_HTTP),
+        stdio: stdio("mcpServers"),
+        http: Some(HttpShape {
+            url_field: None,
+            header_field: None,
+            tag_key: Some("type"),
+            tag_value: Some("http"),
+            sse_tag_value: None,
+            injects: NONE,
+            supports_oauth: true,
+        }),
+        project: Some(ProjectSurface {
+            stdio: stdio("mcpServers"),
+            http: Some(HttpShape {
+                url_field: None,
+                header_field: None,
+                tag_key: Some("type"),
+                tag_value: Some("http"),
+                sse_tag_value: None,
+                injects: NONE,
+                supports_oauth: true,
+            }),
+        }),
+        sources: McpSurfaceSources {
+            first_party: "https://docs.cursor.com/context/model-context-protocol",
+            smithery: Some(SMITHERY_URL),
+            notes: None,
+            verified: VERIFIED,
+        },
+    },
+    skill: skill(
+        paths(
+            &["$HOME/.cursor/skills"],
+            &["$HOME/.cursor/skills"],
+            &["$USERPROFILE\\.cursor\\skills"],
+        ),
+        "https://cursor.com/docs/skills",
+        "2026-07-22",
+    ),
+};
+
+const VSCODE_STDIO: StdioShape = StdioShape {
+    tag_key: Some("type"),
+    tag_value: Some("stdio"),
+    ..stdio("servers")
+};
+const VSCODE_HTTP: HttpShape = HttpShape {
+    url_field: None,
+    header_field: None,
+    tag_key: Some("type"),
+    tag_value: Some("http"),
+    sse_tag_value: None,
+    injects: NONE,
+    supports_oauth: true,
+};
+const VSCODE: HarnessDefinition = HarnessDefinition {
+    id: AgentId::VsCode,
+    display_name: "Visual Studio Code",
+    install_check_paths: paths(
+        &["/Applications/Visual Studio Code.app"],
+        &["$HOME/.config/Code"],
+        &["$APPDATA\\Code"],
+    ),
+    mcp: McpSurface {
+        system_paths: paths(
+            &["$HOME/Library/Application Support/Code/User/mcp.json"],
+            &["$HOME/.config/Code/User/mcp.json"],
+            &["$APPDATA\\Code\\User\\mcp.json"],
+        ),
+        project_file: Some(".vscode/mcp.json"),
+        format: ConfigFormat::Json,
+        system_transports: STDIO_SSE_HTTP,
+        project_transports: Some(STDIO_SSE_HTTP),
+        stdio: VSCODE_STDIO,
+        http: Some(VSCODE_HTTP),
+        project: Some(ProjectSurface {
+            stdio: VSCODE_STDIO,
+            http: Some(VSCODE_HTTP),
+        }),
+        sources: McpSurfaceSources {
+            first_party: "https://code.visualstudio.com/docs/copilot/chat/mcp-servers",
+            smithery: Some(SMITHERY_URL),
+            notes: None,
+            verified: VERIFIED,
+        },
+    },
+    skill: skill(
+        paths(
+            &["$HOME/.copilot/skills"],
+            &["$HOME/.copilot/skills"],
+            &["$USERPROFILE\\.copilot\\skills"],
+        ),
+        "https://code.visualstudio.com/docs/agent-customization/agent-skills",
+        "2026-07-22",
+    ),
+};
+
+const CODEX: HarnessDefinition = HarnessDefinition {
+    id: AgentId::Codex,
+    display_name: "Codex",
+    install_check_paths: paths(
+        &["$HOME/.codex"],
+        &["$HOME/.codex"],
+        &["$USERPROFILE\\.codex"],
+    ),
+    mcp: McpSurface {
+        system_paths: paths(
+            &["$HOME/.codex/config.toml"],
+            &["$HOME/.codex/config.toml"],
+            &["$USERPROFILE\\.codex\\config.toml"],
+        ),
+        project_file: None,
+        format: ConfigFormat::Toml,
+        system_transports: STDIO_HTTP,
+        project_transports: None,
+        stdio: stdio("mcp_servers"),
+        http: Some(HttpShape {
+            url_field: None,
+            header_field: Some("http_headers"),
+            tag_key: None,
+            tag_value: None,
+            sse_tag_value: None,
+            injects: NONE,
+            supports_oauth: true,
+        }),
+        project: None,
+        sources: McpSurfaceSources {
+            first_party: "https://developers.openai.com/codex/mcp",
+            smithery: Some(SMITHERY_URL),
+            notes: Some(
+                "TOML config. Streamable-HTTP entries carry `url`, optional `bearer_token_env_var`, `http_headers`, `env_http_headers`. SSE is not parsed.",
+            ),
+            verified: VERIFIED,
+        },
+    },
+    skill: skill(
+        paths(
+            &["$HOME/.agents/skills"],
+            &["$HOME/.agents/skills"],
+            &["$USERPROFILE\\.agents\\skills"],
+        ),
+        "https://learn.chatgpt.com/docs/build-skills#where-to-save-skills",
+        "2026-07-22",
+    ),
+};
+
+const ZED: HarnessDefinition = HarnessDefinition {
+    id: AgentId::Zed,
+    display_name: "Zed",
+    install_check_paths: paths(
+        &["$HOME/.config/zed"],
+        &["$HOME/.config/zed"],
+        &["$APPDATA\\Zed"],
+    ),
+    mcp: McpSurface {
+        system_paths: paths(
+            &["$HOME/.config/zed/settings.json"],
+            &["$HOME/.config/zed/settings.json"],
+            &["$APPDATA\\Zed\\settings.json"],
+        ),
+        project_file: None,
+        format: ConfigFormat::Json,
+        system_transports: STDIO_SSE_HTTP,
+        project_transports: None,
+        stdio: StdioShape {
+            injects: ZED_INJECTS,
+            ..stdio("context_servers")
+        },
+        http: Some(HttpShape {
+            url_field: None,
+            header_field: None,
+            tag_key: None,
+            tag_value: None,
+            sse_tag_value: None,
+            injects: ZED_INJECTS,
+            supports_oauth: true,
+        }),
+        project: None,
+        sources: McpSurfaceSources {
+            first_party: "https://zed.dev/docs/ai/mcp",
+            smithery: Some(SMITHERY_URL),
+            notes: Some(
+                "Zed calls them \"context_servers\". The block sits alongside other settings, so writes are non-destructive under that key.",
+            ),
+            verified: VERIFIED,
+        },
+    },
+    skill: skill(
+        paths(
+            &["$HOME/.agents/skills"],
+            &["$HOME/.agents/skills"],
+            &["$USERPROFILE\\.agents\\skills"],
+        ),
+        "https://zed.dev/docs/ai/skills",
+        "2026-07-22",
+    ),
+};
+
+const OPENCODE: HarnessDefinition = HarnessDefinition {
+    id: AgentId::OpenCode,
+    display_name: "OpenCode",
+    install_check_paths: paths(
+        &[
+            "$XDG_CONFIG_HOME/opencode",
+            "$HOME/.config/opencode",
+            "$HOME/.opencode",
+            "$HOME/.local/share/opencode",
+        ],
+        &[
+            "$XDG_CONFIG_HOME/opencode",
+            "$HOME/.config/opencode",
+            "$HOME/.opencode",
+            "$HOME/.local/share/opencode",
+        ],
+        &[
+            "$USERPROFILE\\.config\\opencode",
+            "$USERPROFILE\\.opencode",
+            "$USERPROFILE\\.local\\share\\opencode",
+        ],
+    ),
+    mcp: McpSurface {
+        system_paths: paths(
+            &[
+                "$XDG_CONFIG_HOME/opencode/opencode.json",
+                "$HOME/.config/opencode/opencode.json",
+                "$XDG_CONFIG_HOME/opencode/opencode.jsonc",
+                "$HOME/.config/opencode/opencode.jsonc",
+                "$HOME/.opencode/opencode.jsonc",
+            ],
+            &[
+                "$XDG_CONFIG_HOME/opencode/opencode.json",
+                "$HOME/.config/opencode/opencode.json",
+                "$XDG_CONFIG_HOME/opencode/opencode.jsonc",
+                "$HOME/.config/opencode/opencode.jsonc",
+                "$HOME/.opencode/opencode.jsonc",
+            ],
+            &[
+                "$USERPROFILE\\.config\\opencode\\opencode.json",
+                "$USERPROFILE\\.config\\opencode\\opencode.jsonc",
+                "$USERPROFILE\\.opencode\\opencode.jsonc",
+            ],
+        ),
+        project_file: None,
+        format: ConfigFormat::Jsonc,
+        system_transports: STDIO_SSE_HTTP,
+        project_transports: None,
+        stdio: StdioShape {
+            env_field: Some("environment"),
+            command_as_array: true,
+            injects: OPENCODE_STDIO_INJECTS,
+            ..stdio("mcp")
+        },
+        http: Some(HttpShape {
+            url_field: None,
+            header_field: None,
+            tag_key: Some("type"),
+            tag_value: Some("remote"),
+            sse_tag_value: None,
+            injects: OPENCODE_HTTP_INJECTS,
+            supports_oauth: true,
+        }),
+        project: None,
+        sources: McpSurfaceSources {
+            first_party: "https://opencode.ai/docs/mcp",
+            smithery: Some(SMITHERY_URL),
+            notes: Some(
+                "Command written as a single array of [command, ...args] under `command`. `env` renamed to `environment`. Both stdio and remote entries carry a `type` field (`local` vs `remote`).",
+            ),
+            verified: VERIFIED,
+        },
+    },
+    skill: skill(
+        paths(
+            &[
+                "$XDG_CONFIG_HOME/opencode/skills",
+                "$HOME/.config/opencode/skills",
+            ],
+            &[
+                "$XDG_CONFIG_HOME/opencode/skills",
+                "$HOME/.config/opencode/skills",
+            ],
+            &["$USERPROFILE\\.config\\opencode\\skills"],
+        ),
+        "https://opencode.ai/docs/skills",
+        "2026-07-22",
+    ),
+};
+
+const ANTIGRAVITY: HarnessDefinition = HarnessDefinition {
+    id: AgentId::Antigravity,
+    display_name: "Antigravity",
+    install_check_paths: paths(
+        &["$HOME/.gemini/antigravity"],
+        &["$HOME/.gemini/antigravity"],
+        &["$USERPROFILE\\.gemini\\antigravity"],
+    ),
+    mcp: McpSurface {
+        system_paths: paths(
+            &["$HOME/.gemini/config/mcp_config.json"],
+            &["$HOME/.gemini/config/mcp_config.json"],
+            &["$USERPROFILE\\.gemini\\config\\mcp_config.json"],
+        ),
+        project_file: None,
+        format: ConfigFormat::Json,
+        system_transports: STDIO_HTTP,
+        project_transports: None,
+        stdio: stdio("mcpServers"),
+        http: Some(HttpShape {
+            url_field: Some("serverUrl"),
+            header_field: None,
+            tag_key: None,
+            tag_value: None,
+            sse_tag_value: None,
+            injects: NONE,
+            supports_oauth: true,
+        }),
+        project: None,
+        sources: McpSurfaceSources {
+            first_party: "https://antigravity.google/",
+            smithery: Some(SMITHERY_URL),
+            notes: Some(
+                "Google's Antigravity editor. Config lives at `~/.gemini/config/mcp_config.json` (schema id: https://antigravity.google/schemas/mcp_config.json). Uses `serverUrl` for remote entries (matches Windsurf's convention).",
+            ),
+            verified: VERIFIED,
+        },
+    },
+    skill: skill(
+        paths(
+            &["$HOME/.gemini/config/skills"],
+            &["$HOME/.gemini/config/skills"],
+            &["$USERPROFILE\\.gemini\\config\\skills"],
+        ),
+        "https://antigravity.google/docs/skills",
+        "2026-07-22",
+    ),
+};
+
+const CATALOG: [&HarnessDefinition; 7] = [
+    &CLAUDE_CODE,
+    &CODEX,
+    &CURSOR,
+    &OPENCODE,
+    &ANTIGRAVITY,
+    &VSCODE,
+    &ZED,
+];
+
+/// Returns the complete BrowserClaw harness catalog in stable order.
+pub fn list_supported_agents() -> Vec<AgentId> {
+    CATALOG.iter().map(|entry| entry.id).collect()
+}
+
+/// Checks whether a string is one of the seven supported agent identifiers.
+pub fn is_agent_supported(agent: &str) -> bool {
+    CATALOG.iter().any(|entry| entry.id.as_str() == agent)
+}
+
+pub(crate) fn get_catalog_entry(agent: AgentId) -> &'static HarnessDefinition {
+    match agent {
+        AgentId::ClaudeCode => &CLAUDE_CODE,
+        AgentId::Codex => &CODEX,
+        AgentId::Cursor => &CURSOR,
+        AgentId::OpenCode => &OPENCODE,
+        AgentId::Antigravity => &ANTIGRAVITY,
+        AgentId::VsCode => &VSCODE,
+        AgentId::Zed => &ZED,
+    }
+}
+
+/// Resolves the active configuration shape and transport set for an agent.
+pub fn resolve_agent_surface(agent: AgentId, scope: AgentScope) -> Result<AgentSurface, Error> {
+    ensure_system_scope(agent, scope)?;
+    let harness = get_catalog_entry(agent);
+    Ok(AgentSurface {
+        harness,
+        mcp: &harness.mcp,
+        supported_transports: harness.mcp.system_transports,
+        stdio: harness.mcp.stdio,
+        http: harness.mcp.http,
+    })
+}
+
+/// Returns the common harness definition and its separate integration surfaces.
+pub fn resolve_harness_definition(agent: AgentId) -> &'static HarnessDefinition {
+    get_catalog_entry(agent)
+}
+
+/// Resolves the first existing system config candidate, or the first resolvable candidate.
+pub fn resolve_agent_mcp_config_path(agent: AgentId, scope: AgentScope) -> Result<PathBuf, Error> {
+    ensure_system_scope(agent, scope)?;
+    let candidates = selected_os_paths(&get_catalog_entry(agent).mcp.system_paths);
+    if candidates.is_empty() {
+        return Err(Error::UnresolvedConfigPath {
+            agent,
+            reason: format!(
+                "no system config path configured for OS {}",
+                std::env::consts::OS
+            ),
+        });
+    }
+    pick_config_path(candidates)?.ok_or_else(|| Error::UnresolvedConfigPath {
+        agent,
+        reason: "no system config path resolves (env vars unset?)".to_string(),
+    })
+}
+
+pub(crate) fn has_install_fingerprint(agent: AgentId) -> Result<bool, Error> {
+    let checks = selected_os_paths(&get_catalog_entry(agent).install_check_paths);
+    any_exists(checks)
+}
+
+/// Reports catalog install checks separately from config-path writability.
+pub fn detect_installed_agents() -> Result<Vec<AgentInfo>, Error> {
+    CATALOG
+        .iter()
+        .map(|entry| {
+            let installed = has_install_fingerprint(entry.id)?;
+            let config_path = resolve_agent_mcp_config_path(entry.id, AgentScope::System).ok();
+            Ok(AgentInfo {
+                id: entry.id,
+                display_name: entry.display_name.to_string(),
+                config_path,
+                installed,
+            })
+        })
+        .collect()
+}
+
+pub(crate) fn ensure_system_scope(agent: AgentId, scope: AgentScope) -> Result<(), Error> {
+    if scope == AgentScope::System {
+        return Ok(());
+    }
+    Err(Error::UnresolvedConfigPath {
+        agent,
+        reason: "project scope is not supported; only system scope is implemented".to_string(),
+    })
+}
