+++ packages/browseros-agent/crates/harness-integrations/src/lib.rs
@@ -8,13 +8,13 @@ mod planner;
 mod types;
 
 pub use catalog::{
-    AgentSurface, ClientConfig, ClientConfigSources, ConfigFormat, HttpShape, InjectValue,
-    KeyTransform, PerOsPaths, ProjectSurface, StdioShape, detect_installed_agents,
-    is_agent_supported, list_supported_agents, resolve_agent_mcp_config_path,
-    resolve_agent_surface,
+    AgentSurface, ConfigFormat, HarnessDefinition, HttpShape, InjectValue, KeyTransform,
+    McpSurface, McpSurfaceSources, PerOsPaths, ProjectSurface, SkillSurface, StdioShape,
+    detect_installed_agents, is_agent_supported, list_supported_agents,
+    resolve_agent_mcp_config_path, resolve_agent_surface, resolve_harness_definition,
 };
 pub use error::Error;
-pub use manager::Manager;
+pub use manager::McpManager;
 pub use paths::is_installed;
 pub use types::{
     AgentId, AgentInfo, AgentScope, DisconnectInput, DisconnectSummary, LinkInput, LinkSummary,
