+++ packages/browseros-agent/crates/harness-integrations/src/paths.rs
